pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "thermodynamic-diffusion"
version = "0.1.0"
description = "A minimal DDPM implementation on MNIST foregrounding thermodynamic origins"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.8"
dependencies = [
    "torch>=2.0",
    "torchvision",
    "numpy",
    "matplotlib",
    "pyyaml",
    "tqdm",
    "einops",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-cov",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]
